Defer loading of the dev logger to a module
This will end up increasing our code coverage a tiny bit.
