+++ src/notebook/logger.js
@@ -1,0 +1,17 @@
+import Immutable from 'immutable';
+
+const createLogger = require('redux-logger');
+
+export default function clogger() {
+  const logger = createLogger({
+    stateTransformer: (state) =>
+      Object.keys(state).reduce((prev, key) =>
+        Object.assign(
+          {},
+          prev,
+          { [key]: Immutable.Iterable.isIterable(state[key]) ? state[key].toJS() : state[key] }
+        )
+    , {}),
+  });
+  return logger;
+}
